refactor: split tps
