+++ solagent-adapters/rig/solana/src/lib.rs
@@ -17,7 +17,9 @@
 // pub mod deploy_token;
 // pub mod get_balance;
 // pub mod get_balance_other;
+
 pub mod get_tps;
+
 // pub mod get_wallet_address;
 // pub mod mint_nft;
 // pub mod request_faucet_funds;
